Finished SeeSampler and SamplerAgent, collected see corpus, wrote unit tests, updated project status

+++ src/test/kotlin/io/github/omicron2d/communication/messages/IncomingInitMessageTest.kt
@@ -40,7 +40,7 @@ class IncomingInitMessageTest {
 
                 val msg = IncomingInitMessage.deserialise(testInitMessage)
             }
-        } / MESSAGE_DESERIALISATION_COUNT
+        }.toDouble() / MESSAGE_DESERIALISATION_COUNT
 
         // give it 35 ms tolerance (should be fine on most computers)
         assertTrue(time <= MESSAGE_DESERIALISATION_TIME)
